Add dialog for adding new questions
